refactor: move cssVar into its own module and add transitionListener function
